Refactor calls for Xelish Hashing Still a while to go....
