+++ src/hash.h
@@ -362,6 +362,7 @@ extern double algoHashTotal[20];
 extern int algoHashHits[20];
 
 
+#include "util.h"
 template<typename T1>
 inline uint256 pepe_hash(const T1 pbegin, const T1 pend, const uint256 PrevBlockHash)
 {
@@ -375,6 +376,7 @@ inline uint256 pepe_hash(const T1 pbegin, const T1 pend, const uint256 PrevBlock
     static unsigned char pblank[1];
 
     uint512 hash[8];
+
 
     sph_blake512_init(&ctx_blake);
     sph_blake512(&ctx_blake, (pbegin == pend ? pblank : static_cast<const void*>(&pbegin[0])), (pend - pbegin) * sizeof(pbegin[0]));
@@ -412,10 +414,10 @@ inline uint256 pepe_hash(const T1 pbegin, const T1 pend, const uint256 PrevBlock
     return hash[7].trim256();
 }
 
-template<typename T1> inline uint256 xelisv2_hash(const T1 pbegin, const T1 pend, const uint256 PrevBlockHash)
-{   
-	    uint512 hash[8];
-	     return hash[7].trim256();
-}
+// template<typename T1> inline uint256 xelisv2_hash(const T1 pbegin, const T1 pend, const uint256 PrevBlockHash)
+// {   
+	    // uint512 hash[8];
+	     // return hash[7].trim256();
+// }
 
 #endif // RAVEN_HASH_H
