+++ src/pow.cpp
@@ -231,11 +231,13 @@ bool CheckProofOfWork(uint256 hash, unsigned int nBits, const Consensus::Params&
     bool fOverflow;
     arith_uint256 bnTarget;
 
+    LogPrintf("CheckProofOfWork:  Hash is  %s\n",hash.ToString());
     bnTarget.SetCompact(nBits, &fNegative, &fOverflow);
 
     // Check range
     if (fNegative || bnTarget == 0 || fOverflow || bnTarget > UintToArith256(params.powLimit))
         return error("CheckProofOfWork(): nBits below minimum work");
+
 
     // Check proof of work matches claimed amount
     if (UintToArith256(hash) > bnTarget)
